// src/app/Models/RetailerCompanyProfileRequest.ts
export class RetailerCompanyProfileRequest {
    registeredName: string | undefined;
    businessName: string | undefined;
    primaryContactName: string | undefined;
    primaryContactNumber: string | undefined;
    primaryEmailAddress: string | undefined;
    secondaryContactName: string | undefined;
    secondaryContactNumber: string | undefined;
    secondaryEmailAddress: string | undefined;
    pan: string | undefined;
    createdDate: string | undefined;
    createdBy: string | undefined;
    modifiedDate: string | undefined;
    modifiedBy: string | undefined;
}